refactor: migrate datasets for an old running instance (#8)
In this PR, we include the automatic migration workflow.

1. Create the new `ar.datasets` index
2. Reindex documents from the old `.rubrix.datasets-v0` (if any) index into `ar.datasets`
3. Scan every document in `.rubrix.datasets-v0` and create an alias matching the new argilla dataset records index naming convention. (`ar.dataset.<dataset_id>`)

The configured namespace will be used for the migration. So, if a namespace `A` is defined (`ARGILLA_NAMESPACE`), all actions in the migration, will be related to that workspace.

Refs: https://github.com/recognai/Team/issues/42
(cherry picked from commit c2c9191907b72a2df77d52e09edeb223b1a7bd99)

Also:
- Enable migration by environment variable ARGILLA_ENABLE_MIGRATION
- Delete alias when delete a dataset