feat: Unify record-level actions for all tasks (Validate, Clear, and Discard) (#2280)
# Description

This PR aligns record-level actions for all tasks (Validate, Clear, and
Discard) and include the new design

Closes #2266
See #2264

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [x] New feature (non-breaking change which adds functionality)

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

- [x] test for RecordActionButtons
- [x] update snapshot

**Checklist**

- [x] I have merged the original branch into my forked branch
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works