Enhancement: Distinguish between error message and context in validation of spans (#2329)
# Description

I expanded slightly on the error message provided when providing spans
that do not match the tokenization.
Consider the following example script:
```python
import argilla as rg

record = rg.TokenClassificationRecord(
    text = "This is my text",
    tokens=["This", "is", "my", "text"],
    prediction=[("ORG", 0, 6), ("PER", 8, 14)],
)
```
The (truncated) output on the `develop` branch:
```
ValueError: Following entity spans are not aligned with provided tokenization
Spans:
- [This i] defined in This is my ...
- [my tex] defined in ...s is my text
Tokens:
['This', 'is', 'my', 'text']
```
The distinction between `defined in` and `This is` is unclear. I've
worked on this.

The (truncated) output after this PR:
```
ValueError: Following entity spans are not aligned with provided tokenization
Spans:
- [This i] defined in 'This is my ...'
- [my tex] defined in '...s is my text'
Tokens:
['This', 'is', 'my', 'text']
```
Note the additional `'`. Note that the changes rely on `repr`, so if the
snippet contains `'` itself, it uses `"` instead, e.g.:
```python
import argilla as rg

record = rg.TokenClassificationRecord(
    text = "This is Tom's text",
    tokens=["This", "is", "Tom", "'s", "text"],
    prediction=[("ORG", 0, 6), ("PER", 8, 16)],
)
```
```
ValueError: Following entity spans are not aligned with provided tokenization
Spans:
- [This i] defined in 'This is Tom...'
- [Tom's te] defined in "...s is Tom's text"
Tokens:
['This', 'is', 'Tom', "'s", 'text']
```

**Type of change**

- [x] Improvement (change adding some improvement to an existing
functionality)

**How Has This Been Tested**

Modified the relevant tests, ensured they worked.

**Checklist**

- [x] I have merged the original branch into my forked branch
- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I added comments to my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works

- Tom Aarsen

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>