Add timestamps columns and fields to database models (#2378)
This PR include the following changes:
* Add `inserted_at` and `updated_at` columns to all database table
migrations.
* Add `inserted_at` and `updated_at` fields to database models and
Pydantic schemas.

Missing things:
* It was not possible to add `inserted_at`, and `updated_at` fields to
`User` Pydantic schema due to some errors in old code. We can fix this
later.

Gotchas:
* In order to have the same exact values for `inserted_at` and
`updated_at` when a record is inserted I have created a
`default_inserted_at` function that get the context of the operation to
extract the value of `inserted_at` and use it on `update_at`.
* I can see microseconds level date times inserted on SQLite so I guess
we are covered on that sense (maybe we can check this once we test the
changes with PostgreSQL).