Add explicit ascendent order by inserted_at column to collection endpoints (#2397)
This PR add explicit ordering by db insertion time (using `inserted_at`
column) to endpoints that are returning collections from database. Right
now those are for users and workspaces.

This explicit order avoid leaving the order to the database so it is
consistent between SQLite and PostgreSQL or other databases, and always
returning the same result.