Add unique constraint to workspaces name column (#2396)
This PR add an unique constraint to column `name` on `workspaces` table.

This restriction is necessary right now because we are using workspaces
name inside elastic-search indexes and it's not possible to have
duplications. In the future we will have information about elastic
indexes in the database (using ids) so we can rethink is the uniqueness
is necessary.

I have added a missing `mapped_column(unique=True)` for `username` field
on `User` model too.