Add factories and additional tests for new endpoints (#2415)
This PR add tests for new endpoints and also:
- Adding [`factory-boy`](https://factoryboy.readthedocs.io/en/stable/)
as development dependency.
- Adding factories: `UserFactory`, `WorkspaceFactory`, and
`UserWorkspaceFactory`.
- Fixing relationship warnings for `UserWorkspace` SQLAlchemy model.

Additionally some of the tests are marked as pending for two different
reasons:
- SQLite is by default [not enforcing referencial integrity with foreign
keys](https://docs.sqlalchemy.org/en/20/dialects/sqlite.html#foreign-key-support),
we have different options here:
1. Enforce foreign key support executing `PRAGMA` statement on
connection.
  2. Add additional checks to the endpoints where necessary.
- Pydantic is by default accepting empty strings as valid values for
fields that are not optional. This is causing for example that create an
user with an empty string as `first_name` is totally valid.
- We need to check if it's necessary to add a constraint to our models
or an alternative solution (it is a dangerous default behavior).

We have discussed these points and we will solve them as follows:
1. We will enforce SQLite foreign keys integrity using a `PRAGMA`
statement in every connection.
    - The changes are included in this PR.
- Even with this I have added some additional checks in some endpoints
to give more information to the user when necessary.
2. For Pydantic models accepting empty strings for non-optional string
fields:
- We will create a new PR with a solution for this, adding more
constraints of changing this Pydantic default behavior.
3. We will change `users_workspaces` table name and models to
`workspaces_users`, ' WorkspaceUser`, etc. :
     - The changes will be done in a different PR.
     - This has sense because the most important entity is workspace.