Rename users_workspaces to workspaces_users (#2436)
In this PR I'm redefining the order of the entities in
`users_workspaces` table to `workspaces_users`. This will improve the
semantics of our endpoints and make it more clear.

I have changed too the associated ORM models and Pydantic schemas to
reflect this change.

Everything else should remain the same, behavior is not changing.