Feature : data label schema for labels management (#2384)
# Description
Fetch and store dataset labels as proper label schema

Closes #2210 

**Type of change**
- [x] remove obsolete code 
  -  _getDatasetSettings() in Dataset.js
  - initialize() in TextClassification.js and TokenClassification.js
  - initialize() call in the fetchByName() function in Dataset.js
- [x] get labels from api for TextClassification and TokenClassification
- [x] store labels as GlobalEntities in specific vuexorm (see "Ideal
Orm" in this
[link](https://www.notion.so/argilla/Refactoring-60c997e026644a28a2ff6a20c4cd90c3))
- [x] homogenized "AvailableRecords" (from TextClassificationHeader.vue)
and "visibleEntities"(from TokenClassificationHeader) by the same
vuexOrm query
- [x] prepare vuexorm queries to be able to sort globalLabels by any
param (by text, id, ...) and also ascending/descending
- [x] add more feedbacks with toast component if label could be saved,
there was a problem on save or if user try to add a label already saved



Note : 
- With these updates, globalLabels from tokenclassification and
textclassification have the same shape. M**eans they both have a
"color_id" param which can be used to add color for labels in
textClassification**
- the connection between globalEntities and dataset table is with the
"dataset_id" attribute from globalEntities. I didn't implement a
oneToMany relationship to not performance problem as in Weak labelling
implementation. This could be an idea of refactoring for later

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] Refactor (change restructuring the codebase without changing
functionality)
- [x] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

- [x] TokenClassification task
- [x] TextClassification task
- N/A Text2Text task
 
**Checklist**

- [x] I have merged the original branch into my forked branch
- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works

---------

Co-authored-by: keithCuniah <keithcuniah@gmail.comh>