Improve field constraints for some Pydantic schemas (#2432)
We found that Pydantic schema string fields allow empty strings by
default, being the field optional or not.

More info about that here:
https://github.com/pydantic/pydantic/issues/1626

With this behavior we can finish with empty strings used in fields like
`user.first_name`, `user.last_name` and others. Something that can have
a bad impact in our application.

I have explored different solutions, like defining a common validation
or a custom field type. At the end I think the most simple solution is
to be explicit and define the constraint that we want for the fields. In
this case `min_length=1` when necessary.

Additional changes:
* Added new max length constraint for `user.password` field.
* Notice how I have changed `constr` for `str = Field(...)`. If I'm not
wrong looks like they are similar but `Field` is used to generate a
validation response when errors are found.