refactor: Improve efficiency of  `.scan` (and `.load`) if `limit` is set (#2434)
Hello!

## Pull Request overview
* Refactor `Datasets.scan` to never request more samples than required.
* Implements a `batch_size` variable which is now set to
`self.DEFAULT_SCAN_SIZE`. This parameter can easily be exposed in the
future for `batch_size` support, as discussed previously in Slack etc.

## Details
The core of the changes is that the URL is now built on-the-fly at every
request, rather than just once. For each request, the limit for that
request is computed using `min(limit, batch_size)`, and `limit` is
decremented, allowing it to represent a "remaining samples to fetch".
When `limit` reaches 0, i.e. 0 more samples to fetch, we return from
`scan`.

I've also added a check to ensure that the `limit` passed to `scan`
can't negative.

## Tests
For tests, I've created a `test_scan_efficient_limiting` function which
verifies the new and improved behaviour. It contains two monkeypatches:
1. `DEFAULT_SCAN_SIZE` is set to 10. Because our dataset in this test
only has 100 samples, we want to ensure that we can't just sample the
entire dataset with one request.
2. The `http_client.post` method is monkeypatched to allow us to track
the calls to the server.

We test the following scenarios:
* `limit=23` -> 3 requests: for 10, 10 and 3 samples.
* `limit=20` -> 2 requests: for 10, 10 samples.
* `limit=6` -> 1 request: for 6 samples.

There's also a test to cover the new ValueError if `limit` < 0.

## Effects
Consider the following script:
```python
import argilla as rg
import time

def test_scan_records(fields):
    client = rg.active_client()

    data = client.datasets.scan(
        name="banking77-topics-setfit",
        projection=fields,
        limit=1,  # <- Note, just one sample
    )
    start_t = time.time()
    list(data)
    print(f"load time for 1 sample with fields={fields}: {time.time() - start_t:.4f}s")

test_scan_records(set())
test_scan_records({"text"})
test_scan_records({"tokens"})
```

On this PR, this outputs:
```
load time for 1 sample with fields=set(): 0.0774s
load time for 1 sample with fields={'text'}: 0.0646s
load time for 1 sample with fields={'tokens'}: 0.0669s
```
On the `develop` branch, this outputs:
```
load time for 1 sample with fields=set(): 0.1355s
load time for 1 sample with fields={'text'}: 0.1347s
load time for 1 sample with fields={'tokens'}: 0.1173s
```
This can be repeated for `rg.load(..., limit=1)`, as that relies on
`.scan` under the hood.

Note that this is the most extreme example of performance gains. The
performance increases in almost all scenarios if `limit` is set, but the
effects are generally not this big. Going from fetching 250 samples 8
times to fetching 250 samples 7 times and 173 once doesn't have as big
of an impact.

---

**Type of change**

- [x] Refactor (change restructuring the codebase without changing
functionality)

**How Has This Been Tested**

See above.

**Checklist**

- [x] I have merged the original branch into my forked branch
- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works

---

- Tom Aarsen