Add basic authorization policies (#2491)
This is a first proposal of the new authorization policies after
visiting them for the first time on @frascuchon PR:
https://github.com/argilla-io/argilla/pull/2469

The general idea is to have a `policies` package where we can create
different policies classes including the authorization logic for every
model in our application.

The `policies` package will include also two method to facilitate check
authorization policies on API handlers.
- `authorized`: that will raise a `ForbiddenOperationError` exception if
the policy does not check.
- `is_authorized`: that will return `True` or `False` depending if the
policy checks or not.

Things to have into account:
- The user that we get as `current_user` on API handlers is not an
SQLAlchemy ORM `User` model (we need to take a look to this
@frascuchon).
- The creation of default user on server startup it's breaking some of
my tests so I have disabled it (we can discuss if with the new user
creation task this can be removed).
- There is an alternative version of authorization policies that we can
implement using closures. It has a lot of flexibility but adds a little
bit of complexity to the code.