Add argilla_data volume to docker compose (#2560)
In this PR we are defining an argilla data volume to be used on our
`docker-compose.yaml`.

Things to have into account:
- <del>I'm defining the volume as `data` because docker adds the
`argilla_` prefix once it's created being `argilla_data` at the
end.</del>
- Environment variables defined on the docker compose cannot being used
on the right side of the volumes section. This can potentially provoke
errors if the users sets the `ARGILLA_HOME_PATH` value.

I have successfully tested to set a custom volume pointing to a host
folder with:

```dockerfile
argilla:
    volumes:
      - ${HOME}/.argilla:/var/lib/argilla
     
networks:
  argilla:
    driver: bridge
volumes:
  # argilladata:
  elasticdata:
```

And the database file is correctly created on my home folder. Restarting
the docker compose is working too.