Add new ADMIN_ENABLED environment variable to release.Dockerfile (#2601)
This PR adds a new `ADMIN_ENABLED` environment variable with a default
value of `true`. This environment variable is checked to create or not
the default admin user for Docker image generated by
`release.Dockerfile`.

We will set this environment variable to `false` later for `PRE` and
`PRO` Docker compose file so the default admin user is not created on
these environments.