fix: Using backend config to check version (#2597)
The backend version check request is not using the backend configuration
setup from Argilla, so in some security environments, this check may
fail.

Instead of using a standard http request, on this PR we use a widely
compatible `OpenSearch` client to request the version info.

Closes #2311

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [x] Bug fix (non-breaking change which fixes an issue)

**How Has This Been Tested**

Manual test in a secured environment

**Checklist**

- [x] I have merged the original branch into my forked branch
- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works