Create es index when a dataset published (#2700)
A basic integration with elasticsearch when a dataset is published. At
this point, the elasticsearch index will be created based on dataset
configuration.

A DB rollback is executed if some problem is found creating the index.
If the error is raised from the commit statement, index creation should
be reverted, or another mechanism to recreate the index should be
provided through the API endpoint (and extra param with `force_recreate`
or just `force`). But this won't be tackled in this PR.

Edit: Also, the elasticsearch client deps are reviewed. Now, the named
8x version is used. This allows us more control over the `es` client
version to use, and it can suppress installation conflicts in an
environment with a preview es client package installed.