Ignore errors on index creation (#2729)
This PR changes the default behavior when publishing a draft dataset,
and is a temporal solution. When the search index creation raises some
error, the dataset will be published and a warning will be logged into
the server errors.

Once we have compatibility with our dev environments (some of them using
OpenSearch) we can proceed to change this behavior.