Fix delete datasets with cascade deletion (#2799)
# Description

There was a bug on the backend detected by @keithCuniah when a dataset
with related questions is deleted giving a database error.

This was caused by SQLAlchemy default behavior trying to delete
explicitly relationships rows one by one.

This default behavior can be changed setting `passive_deletes=True` on
the defined model relationship. Meaning that the deletion will not be
done and it will be managed by the database (if available).