Use `/me` prefix for endpoints returning current user contextual response (#2826)
# Description

This PR include changes adding prefix `/me` to APIv1 endpoints that
build the response using current user as context. The behavior of these
endpoints has been changed too.

The include changes are:
* `GET` `/api/v1/datasets` changed to `GET` `/api/v1/me/datasets`
  *  The behavior of the endpoint didn't change.
* `GET` `/api/v1/datasets/{dataset_id}/records` changed to `GET`
`/api/v1/me/datasets/{dataset_id}/records`
  * The behavior for the endpoint returning record didn't change.
* The behavior including record responses changed and now it's only
including current user record responses (even if you are an admin).
* `GET` `/api/v1/records/{record_id}/responses` has been removed.
  * It wasn't being used by the frontend so I've deleted it for now.
 
@keithCuniah feel free to add the necessary frontend changes to this PR
replacing the use of the old endpoints to the new ones according to the
previous changes section. Take into account that endpoint
`/api/v1/me/datasets/{dataset_id}/records` is now only returning the
responses for the current user when `include=response` is used. (Prior
to this change the responses for all users where included when using
`include=response` and the user doing the request was an admin).

Closes #2823

---------

Co-authored-by: Keith Cuniah <88380932+keithCuniah@users.noreply.github.com>
Co-authored-by: keithCuniah <keithcuniah@gmail.comh>