UI: Include status filter for feedback task (#2828)
# Description

This PR includes a new component in the filter area, a status selector
to allow filtering of records in the new Feedback Task.

Related [#2743](https://github.com/argilla-io/argilla/issues/2743)

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)
- [x] New feature (non-breaking change which adds functionality)

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

- [ ] Test A
- [ ] Test B

**Checklist**

- [x] I have merged the original branch into my forked branch
- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)