feat: enable markdown configuration for text fields and questions (#3000)
# Description

This PR adds a param `use_markdown` to the `TextFieldSettings` and
`TextQuestionSettings` classes in API to enable markdown support. By
default, the parameter is disabled, which is compatible with the
previous behavior.


Also, some tests creating dataset fields and questions have been grouped
using the `pytest.mark.parameterize` decorator, since the internal
behavior was the same.

Refs #2998 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [x] Improvement (change adding some improvement to an existing
functionality)

**How Has This Been Tested**

New tests have been created 


**Checklist**

- [x] I have merged the original branch into my forked branch
- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)

---------

Co-authored-by: gabrielmbmb <gmartinbdev@gmail.com>