feat: add `SearchEngine` pagination offset (#3088)
# Description

This PR updates the `SearchEngine.search` method to add an `offset`
parameter (Qdrant, Weaviate, Milvus, etc use `offset` and `limit`
combination too) that will allow to skip results. Also, the `offset`
parameter has been added to the search endpoint as a query parameter.

In addition, this PR fixes an issue for the search endpoint which was
returning an empty list of `responses` for each record even if the
`responses` were not in the `includes` query parameter.

Closes #3074

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)

**How Has This Been Tested**

Manual testing and I've added unit tests.

**Checklist**

- [x] I have merged the original branch into my forked branch
- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)

---------

Co-authored-by: Francisco Aranda <francis@argilla.io>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>