fix: unhandled `ImportError` on `huggingface_hub` (#3174)
# Description

As reported by @dvsrepo, when installing `argilla` without any extra
dependency, an unhandled `ImportError` is raised. That's due to the
import of `ArgillaDatasetCard` being on top of the file instead of under
the `@requires_version` decorator, so that the `ImportError` will still
be raised, but at least it will be handled.

Also the `ImportError` was being triggered just when importing the
`FeedbackDataset` class as it was on top of the file, which was
obviously wrong, as it should just raise when calling
`push_to_huggingface`.

**Type of change**

- [X] Bug fix (non-breaking change which fixes an issue)

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

- [X] Re-run unit tests locally on Python 3.11
- [X] Make sure that `from argilla.client.feedback.dataset import
FeedbackDataset` doesn't raise an `ImportError` when `huggingface_hub`
is not installed

**Checklist**

- [X] I have merged the original branch into my forked branch
- [ ] I added relevant documentation
- [X] follows the style guidelines of this project
- [X] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [X] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)